refactor: consolidate WI15 script scaffolding, drop dead code
Move the three operator scripts' duplicated plumbing (parser skeleton,
required-value/confirmation prompts, run-id stamp, best-effort close,
bench-cleanup fold-in, and the 113-line report finalization block) into
omotion.calibration.script_support. Factories and helper names stay as
per-script module globals so test monkeypatching is unchanged.

Dead code removed: select_closest_valid_setting (both workflows use the
target-parameterized variant), the stage(**data) parameter no call site
used, and the two-object DEFAULT_USER_CONFIG compatibility dance (now a
single immutable mapping plus the copying accessor). Tests pinning the
removed code are deleted with it.
