+++ scripts/wi15_single_sensor_laser_calibration.py
@@ -3,28 +3,33 @@
 from __future__ import annotations
 
 import argparse
-from dataclasses import dataclass, replace
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Callable, Sequence
 
 import omotion
-from omotion.calibration.laser import FailureKind, ProcedureStatus
 from omotion.calibration.laser_hardware import (
     MotionLaserCalibrationBench,
     OphirEnergyMeter,
 )
 from omotion.calibration.reporting import HtmlRunReport, JsonRunRecorder
+from omotion.calibration.script_support import (
+    APPROVED_PROCEDURE_REVISION,
+    PROCEDURE_ID,
+    OperatorCanceled as _OperatorCanceled,
+    OperatorRunReportRequest,
+    close_best_effort as _close_best_effort,
+    confirmed as _confirmed,
+    finalize_run_artifacts,
+    make_parser,
+    required_value as _required_value,
+    utc_run_id as _run_id,
+)
 from omotion.calibration.single_sensor_laser import (
     SingleSensorLaserCalibrationRequest,
-    ReportArtifactEvidence,
-    ReportArtifactStatus,
     SingleSensorLaserCalibrationWorkflow,
 )
 
-
-PROCEDURE_ID = "WI-00015"
-APPROVED_PROCEDURE_REVISION = "WI-00015 automated process addendum approved 2026-08-12"
 
 recorder_factory = JsonRunRecorder
 meter_factory = OphirEnergyMeter
@@ -33,71 +38,22 @@ workflow_factory = SingleSensorLaserCalibrationWorkflow
 report_factory = HtmlRunReport
 
 
-@dataclass(frozen=True)
-class OperatorRunReportRequest:
-    """Report metadata, including the approved procedure revision."""
-
-    request: SingleSensorLaserCalibrationRequest
-    procedure_revision: str
-
-
-class _OperatorCanceled(Exception):
-    pass
-
-
 def _parser() -> argparse.ArgumentParser:
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("--output-dir", default="./wi15_out")
-    parser.add_argument("--operator")
-    parser.add_argument("--build-revision")
-    parser.add_argument("--fixture-id")
-    parser.add_argument("--fixture-calibration-status")
-    parser.add_argument(
-        "--procedure-revision", default=APPROVED_PROCEDURE_REVISION
+    return make_parser(
+        __doc__,
+        lambda parser: parser.add_argument("--fixture-calibration-status"),
     )
-    return parser
 
 
-def _required_value(
-    value: str | None, prompt: str, input_func: Callable[[str], str]
+def _selected_side(
+    input_func: Callable[[str], str], output_func: Callable[[str], None]
 ) -> str:
-    while True:
-        candidate = value if value is not None else input_func(prompt)
-        value = None
-        candidate = candidate.strip()
-        if candidate:
-            return candidate
-
-
-def _confirmed(prompt: str, input_func: Callable[[str], str]) -> bool:
-    response = input_func(prompt).strip().lower()
-    if response in ("yes", "y"):
-        return True
-    if response in ("no", "n", "cancel", "c"):
-        return False
-    return False
-
-
-def _selected_side(input_func: Callable[[str], str], output_func: Callable[[str], None]) -> str:
     while True:
         side = input_func("Installed sensor side (left/right): ").strip().lower()
         if side in ("left", "right"):
             output_func(f"Selected sensor side: {side}")
             return side
         output_func("Enter exactly left or right.")
-
-
-def _run_id() -> str:
-    return datetime.now(timezone.utc).strftime("%Y%m%dT%H%M%SZ")
-
-
-def _close_best_effort(resource) -> None:
-    if resource is None:
-        return
-    try:
-        resource.close()
-    except Exception:
-        pass
 
 
 def main(
@@ -162,73 +118,14 @@ def main(
             started_at=datetime.now(timezone.utc),
         )
         result = workflow.run(request)
-        report_path = Path(recorder.run_directory) / "report.html"
-        incomplete_result = replace(
-            result,
-            report_paths=(Path(recorder.json_path),),
-            report_artifact=ReportArtifactEvidence(
-                report_path, ReportArtifactStatus.INCOMPLETE
-            ),
+        return finalize_run_artifacts(
+            request=request,
+            result=result,
+            recorder=recorder,
+            report_factory=report_factory,
+            procedure_revision=procedure_revision,
+            output_func=output_func,
         )
-        recorder.checkpoint(incomplete_result)
-        try:
-            report = report_factory(recorder.run_directory)
-            report_path = Path(report.report_path)
-            finalized_result = replace(
-                incomplete_result,
-                report_paths=(Path(recorder.json_path), report_path),
-                report_artifact=ReportArtifactEvidence(
-                    report_path, ReportArtifactStatus.FINALIZED
-                ),
-            )
-            written_report = report.write(
-                OperatorRunReportRequest(request, procedure_revision),
-                finalized_result,
-                recorder.json_path,
-            )
-            if (
-                Path(written_report).resolve() != report_path.resolve()
-                or not report_path.is_file()
-            ):
-                raise RuntimeError("HTML report writer did not create the expected file")
-        except Exception as exc:
-            report_failure = str(exc) or exc.__class__.__name__
-            failed_result = replace(
-                incomplete_result,
-                status=(
-                    ProcedureStatus.FAILED
-                    if result.status is ProcedureStatus.PASSED
-                    else result.status
-                ),
-                failure_kind=(
-                    FailureKind.REPORT
-                    if result.status is ProcedureStatus.PASSED
-                    else result.failure_kind
-                ),
-                failure_reason=(
-                    "HTML report generation failed."
-                    if result.status is ProcedureStatus.PASSED
-                    else result.failure_reason
-                ),
-                report_paths=(Path(recorder.json_path),),
-                report_artifact=ReportArtifactEvidence(
-                    Path(report_path),
-                    ReportArtifactStatus.FAILED,
-                    report_failure,
-                ),
-            )
-            recorder.checkpoint(failed_result)
-            output_func(f"HTML report generation failed: {report_failure}")
-            return 1
-        recorder.checkpoint(finalized_result)
-        output_func(f"Terminal status: {finalized_result.status.value}")
-        if finalized_result.failure_kind is not None:
-            output_func(f"Failure category: {finalized_result.failure_kind.value}")
-        if finalized_result.failure_reason is not None:
-            output_func(f"Failure reason: {finalized_result.failure_reason}")
-        output_func(f"JSON evidence: {Path(recorder.json_path).resolve()}")
-        output_func(f"HTML report: {Path(report_path).resolve()}")
-        return 0 if finalized_result.status is ProcedureStatus.PASSED else 1
     except Exception as exc:
         output_func(f"Calibration failed before a terminal report: {exc}")
         return 1
